Better container management and download progress

+++ AzureStorage/CloudAccess.cs
@@ -28,36 +28,32 @@ namespace AzureStorage
                 return blobClient;
             }
         }
-        public static string ContainerName = "root";
-
-        private static CloudBlobContainer RootContainer
-        {
-            get { return BlobClient.GetRootContainerReference(); }
-        }
-        private static CloudBlobContainer Container
-        {
-            get
-            {
-                var container = ContainerName == "root" ? RootContainer : BlobClient.GetContainerReference(ContainerName);
-                if (!container.Exists())
-                {
-                    container = CreateContainer(ContainerName);
-                }
-                container.SetPermissions(new BlobContainerPermissions { PublicAccess = BlobContainerPublicAccessType.Blob });
-                return container;
-            }
-        }
         private const int BlockSize = 256 * 1024;
 
-        public static IEnumerable<IListBlobItem> GetFiles()
+        public static IEnumerable<IListBlobItem> GetFiles(string container)
         {
-            return Container.ListBlobs();
+            return GetContainer(container).ListBlobs();
+        }
+
+        public static CloudBlobContainer GetContainer(string containerName)
+        {
+            return BlobClient.GetContainerReference(containerName);
         }
 
         public static CloudBlobContainer CreateContainer(string containerName)
         {
-            var newContainer = BlobClient.GetContainerReference(containerName);
+            var newContainer = BlobClient.GetContainerReference(containerName.ToLower().Replace(" ",""));
+            
+            if (newContainer.Exists())
+            {
+                newContainer.Metadata["Title"] = containerName;
+                newContainer.SetMetadata();
+                return newContainer;
+            }
             newContainer.CreateIfNotExists();
+            newContainer.Metadata.Add("Title", containerName);
+            newContainer.SetMetadata();
+            newContainer.FetchAttributes();
             return newContainer;
         }
 
@@ -66,13 +62,14 @@ namespace AzureStorage
             BlobClient.GetContainerReference(containerName).DeleteIfExists();
         }
 
-        public static List<string> GetContainers()
+        public static List<CloudBlobContainer> GetContainers()
         {
-            return BlobClient.ListContainers().Select(c => c.Name).ToList();
+            return BlobClient.ListContainers().ToList();
         }
 
-        public static async Task<bool> UploadFile(string fileName, string filePath, int totalFiles, int currentFile, Main mainScreen)
+        public static async Task<bool> UploadFile(string fileName, string filePath, int totalFiles, int currentFile, string container, Main mainScreen)
         {
+            var Container = BlobClient.GetContainerReference(container);
             var blockBlob = Container.GetBlockBlobReference(fileName);
             blockBlob.StreamWriteSizeInBytes = BlockSize;
             var bytesToUpload = (new FileInfo(filePath)).Length;
@@ -123,51 +120,95 @@ namespace AzureStorage
             return true;
         }
 
-        public static async void UploadFiles(string[] fileNames, string[] filePaths, Main mainScreen)
+        public static async void UploadFiles(string[] fileNames, string[] filePaths, bool deleteLocal, string container, Main mainScreen)
         {
             for (var i = 0; i < fileNames.Length; i++)
             {
                 // upload the file
-                await UploadFile(fileNames[i], filePaths[i], fileNames.Length, i + 1, mainScreen);
-                File.Delete(filePaths[i]);
+                await UploadFile(fileNames[i], filePaths[i], fileNames.Length, i + 1, container, mainScreen);
+                if(deleteLocal) File.Delete(filePaths[i]);
                 // update the mainscreen to clear progress bar
                 mainScreen.UpdateProgressBar(0);
                 mainScreen.UpdateStatusText(string.Format("{0} was uploaded successfully.", fileNames[i]));
-                mainScreen.UpdateBlobList();
+                mainScreen.UpdateBlobList(container);
             }
         }
 
-        public static async Task<bool> DownloadFile(string fileName, string filePath, int totalFiles, int currentFile, Main mainScreen)
+        public static async Task<bool> DownloadFile(string fileName, string filePath, int totalFiles, int currentFile, string container, Main mainScreen)
         {
-            var blockBlob = Container.GetBlockBlobReference(fileName);
-            mainScreen.UpdateStatusText(string.Format("{0} Downloading {1}...", totalFiles > 1 ? string.Format("[{0}/{1}]", currentFile, totalFiles) : "", fileName));
-            await blockBlob.DownloadToFileAsync(filePath, FileMode.Create);
+            var blockBlob = GetContainer(container).GetBlobReferenceFromServer(fileName);
+            var bytesToDownload = blockBlob.Properties.Length;
+            long fileSize = bytesToDownload;
+            if (fileSize <= BlockSize)
+            {
+                mainScreen.UpdateStatusText(string.Format("{0} Downloading {1}...", totalFiles > 1 ? string.Format("[{0}/{1}]", currentFile, totalFiles) : "", fileName));
+                await blockBlob.DownloadToFileAsync(filePath, FileMode.Create);
+            }
+            else
+            {
+                int index = 1;
+                long startPosition = 0;
+                long bytesDownloaded = 0;
 
+                // upload bytesToRead bytes at a time
+                if(File.Exists(filePath + ".tmp")) File.Delete(filePath + ".tmp");
+                var fileStream = new FileStream(filePath + ".tmp", FileMode.Create);
+                do
+                {
+                    var bytesToWrite = Math.Min(BlockSize, bytesToDownload);
+                    var blobContents = new byte[bytesToWrite];
+                    var mre = new ManualResetEvent(false);
+                    // download current bytesToWrite bytes
+                    var ado = blockBlob.DownloadRangeToByteArrayAsync(blobContents, 0, startPosition,
+                        (int) bytesToWrite);
+                    await ado.ContinueWith(t =>
+                    {
+                        fileStream.Position = startPosition;
+                        fileStream.Write(blobContents, 0, (int) bytesToWrite);
+                        mre.Set();
+                        bytesDownloaded += bytesToWrite;
+                        bytesToDownload -= bytesToWrite;
+                        startPosition += bytesToWrite;
+                        index++;
+                        var percentComplete = (int) Math.Min((double) bytesDownloaded/fileSize*100, 100);
+                        mainScreen.UpdateProgressBar(percentComplete);
+                        mainScreen.UpdateStatusText(string.Format("{0} ({1}%) Downloading {2}...",
+                            totalFiles > 1 ? string.Format("[{0}/{1}]", currentFile, totalFiles) : "",
+                            percentComplete, fileName));
+
+                    });
+                    mre.WaitOne();
+                } while (bytesToDownload > 0);
+                fileStream.Close();
+            }
             return true;
         }
 
-        public static async void DownloadFiles(string[] fileNames, string[] filePaths, Main mainScreen)
+        public static async void DownloadFiles(string[] fileNames, string[] filePaths, string container, Main mainScreen)
         {
             for (var i = 0; i < fileNames.Length; i++)
             {
-                await DownloadFile(fileNames[i], filePaths[i], fileNames.Length, i + 1, mainScreen);
+                await DownloadFile(fileNames[i], filePaths[i], fileNames.Length, i + 1, container, mainScreen);
+                if (File.Exists(filePaths[i])) File.Delete(filePaths[i]);
+                File.Move(filePaths[i] + ".tmp", filePaths[i]);
+                if (File.Exists(filePaths[i] + ".tmp")) File.Delete(filePaths[i] + ".tmp");
 
                 // update the mainscreen to clear progress bar
                 mainScreen.UpdateProgressBar(0);
                 mainScreen.UpdateStatusText(string.Format("{0} was downloaded successfully.", fileNames[i]));
-                mainScreen.UpdateBlobList();
+                mainScreen.UpdateBlobList(container);
             }
         }
 
-        public static bool DeleteFile(string fileName)
+        public static bool DeleteFile(string fileName, string container)
         {
-            var blockBlob = Container.GetBlockBlobReference(fileName);
+            var blockBlob = GetContainer(container).GetBlockBlobReference(fileName);
             return blockBlob.DeleteIfExists();
         }
 
-        public static bool DeleteFiles(List<string> fileNames)
+        public static bool DeleteFiles(List<string> fileNames, string container)
         {
-            return fileNames.Aggregate(true, (current, deleteFile) => DeleteFile(deleteFile) && current);
+            return fileNames.Aggregate(true, (current, deleteFile) => DeleteFile(deleteFile, container) && current);
         }
     }
 }
